Move playground view module to its own folder
